Move OnDuplicate enum to Destinations namespace (co-locate with consumer)
